requirements.txt
torch==2.2.2
torchvision==0.17.2
pandas>=1.4
scikit-learn>=1.0
matplotlib>=3.5
tqdm>=4.64
omegaconf>=2.3.0
PyYAML>=6.0
tensorboard>=2.10
einops>=0.6.0
pytorch_lightning